(#5) Make Hoard aware of CRUD + a model's collection
When Hoard checks for a model's cache existence, and that model has a
collection, Hoard will check that collection's cache for the model.
Similarly, updates and deletes to a model will update the collection's
cache. The implementation currently assumes that a model in a collection
will continue to be accessed via that collection. This results in only
the collection being cached, so the model will have to create a new
cached copy if it is removed from the collection. This should probably
never happen in practice.